pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "predictive-modeling-ai"
version = "0.1.0"
description = "Ingest public datasets and visualize historic and forecasted trends"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Data
    "pandas==2.2.3",
    "numpy==1.26.4",
    "pyarrow==16.1.0",
    # ML / forecasting
    "scikit-learn==1.5.1",
    "statsmodels==0.14.2",
    "prophet==1.1.5",
    # Visualisation
    "plotly==5.22.0",
    "kaleido==0.2.1",          # static image export for plotly
    # Config / validation
    "pydantic==2.7.4",
    "pydantic-settings==2.3.4",
    # CLI
    "typer[all]==0.12.3",
    # Serving
    "fastapi==0.111.1",
    "uvicorn[standard]==0.30.1",
    # Logging
    "loguru==0.7.2",
    # HTTP (dataset download)
    "httpx==0.27.0",
    # Config parsing
    "pyyaml==6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.2",
    "pytest-cov==5.0.0",
    "ruff==0.4.10",
    "mypy==1.10.1",
    "types-requests==2.32.0.20240712",
    "pre-commit==3.7.1",
]

[project.scripts]
pmai = "main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["data_ingestion*", "features*", "training*", "models*", "service*", "monitoring*", "config*", "utils*"]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["data_ingestion", "features", "training", "models", "service", "monitoring", "config", "utils"]

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=. --cov-report=term-missing --cov-omit=tests/*"
log_cli = true
log_cli_level = "INFO"
